refactor: Move modals back to the modals folder
